app/talk_to_graph/demo/backend/requirements.txt
Flask
Flask-Cors
rdflib
requests
openai
python-dotenv
sentence-transformers
chromadb